Move more types from ToSseEvent to projection actor
Preparation for the final stage of the refactor.
