+++ daemon/src/routes_maker.rs
@@ -3,9 +3,9 @@ use bdk::bitcoin::Network;
 use daemon::auth::Authenticated;
 use daemon::model::cfd::{OrderId, Role};
 use daemon::model::{Price, Usd, WalletInfo};
-use daemon::projection::Feeds;
+use daemon::projection::{CfdAction, Feeds};
 use daemon::routes::EmbeddedFileExt;
-use daemon::to_sse_event::{CfdAction, CfdsWithAuxData, ToSseEvent};
+use daemon::to_sse_event::{CfdsWithAuxData, ToSseEvent};
 use daemon::{maker_cfd, maker_inc_connections, monitor, oracle, wallet};
 use http_api_problem::{HttpApiProblem, StatusCode};
 use rocket::http::{ContentType, Header, Status};
