+++ daemon/src/routes_taker.rs
@@ -2,9 +2,9 @@ use bdk::bitcoin::{Amount, Network};
 use daemon::connection::ConnectionStatus;
 use daemon::model::cfd::{calculate_long_margin, OrderId, Role};
 use daemon::model::{Leverage, Price, Usd, WalletInfo};
-use daemon::projection::Feeds;
+use daemon::projection::{CfdAction, Feeds};
 use daemon::routes::EmbeddedFileExt;
-use daemon::to_sse_event::{CfdAction, CfdsWithAuxData, ToSseEvent};
+use daemon::to_sse_event::{CfdsWithAuxData, ToSseEvent};
 use daemon::{bitmex_price_feed, monitor, oracle, taker_cfd, wallet};
 use http_api_problem::{HttpApiProblem, StatusCode};
 use rocket::http::{ContentType, Status};
